+++ src/ranges/trie_test.go
@@ -2,7 +2,7 @@
 // Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
 // the U.S. Government retains certain rights in this software.
 
-package minicli
+package ranges
 
 import "testing"
